+++ api/src/main/java/com/motogest/api/layers/services/ModelosService.java
@@ -8,15 +8,25 @@ import org.springframework.stereotype.Service;
 
 import com.motogest.api.layers.entities.Marcas;
 import com.motogest.api.layers.entities.Modelos;
+import com.motogest.api.layers.entities.Motos;
+import com.motogest.api.layers.entities.VariacoesModelos;
 import com.motogest.api.layers.repositories.ModelosRepository;
+import com.motogest.api.layers.repositories.MotosRepository;
+import com.motogest.api.layers.repositories.VariacoesModeloRepository;
 
 @Service
 public class ModelosService {
     @Autowired
-    private ModelosRepository modelosRepository;
+    ModelosRepository modelosRepository;
 
     @Autowired
     MarcasService marcasService;
+
+    @Autowired
+    VariacoesModeloRepository variacoesModeloRepository;
+
+    @Autowired
+    MotosRepository motosRepository;
 
     String naoEncontrado = "Não foi possível encontrar um modelo com o parâmetro informado";
 
@@ -91,6 +101,18 @@ public class ModelosService {
     public String deletarModelo(Integer id) {
         Modelos modelo = modelosRepository.findById(id).orElseThrow(() -> new NoSuchElementException(naoEncontrado));
 
+        List<Motos> motos = motosRepository.findByModelo(modelo);
+
+        if (!motos.isEmpty()) {
+            throw new IllegalArgumentException("Não é possível deletar um modelo que possui motos associadas");
+        }
+
+        List<VariacoesModelos> variacoes = variacoesModeloRepository.findByModeloId(modelo.getId());
+
+        if (!variacoes.isEmpty()) {
+            throw new IllegalArgumentException("Não é possível deletar um modelo que possui variações associadas");
+        }
+        
         modelosRepository.delete(modelo);
 
         String feedback = String.format("Modelo '%s' deletado com sucesso!", modelo.getNome());
